Move message processing to separate module, add more filter rules
