+++ signal_scanner_bot/signal_scanner_bot.py
@@ -1,31 +1,13 @@
 """Main module."""
 import logging
 import subprocess
-from datetime import datetime
-from typing import Dict
 
 import ujson
 
 from . import env
-
+from . import messages
 
 log = logging.getLogger(__name__)
-
-
-def process_message(blob: Dict) -> None:
-    log.debug(f"Got message: {blob}")
-    envelope = blob.get("envelope", {})
-    if not envelope or "dataMessage" not in envelope:
-        log.error(f"Malformed message: {blob}")
-
-    data = envelope["dataMessage"]
-    if data is None or not data.get("message"):
-        # No actual message contents
-        return
-
-    message = data["message"]
-    timestamp = datetime.fromtimestamp(data["timestamp"] / 1000.0)
-    log.info(f"{timestamp.isoformat()}: '{message}'")
 
 
 def listen_and_print():
@@ -38,7 +20,11 @@ def listen_and_print():
         for line in iter(proc.stdout.readline, b""):
             line = line.decode("utf-8").rstrip()
             blob = ujson.loads(line)
-            process_message(blob)
+            try:
+                messages.process_message(blob)
+            except Exception:
+                log.error(f"Malformed message: {blob}")
+                raise
         # Check to see if there's any content in stderr
         if proc.stderr.peek(10):
             for line in iter(proc.stderr.readline, b""):
